+++ linux-runner.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# Adiciona Poetry ao PATH para a sessão atual
+export PATH="$HOME/.local/bin:$PATH"
+
+# Executa o comando do Poetry
+poetry run gestao_filiados/manage.py runserver